MCC-SimpleGAN/requirements.txt
numpy==1.18.1
Pillow==6.1.0
torch==1.4.0
torchvision==0.5.0

